ai_agent/requirements.txt
requests
beautifulsoup4
python-dotenv
g4f
google-generativeai
